Renamed categories.py into category.py to remain consistent with naming convention
