+++ blackbook/models/__init__.py
@@ -3,4 +3,4 @@ from .base import get_currency_choices, get_default_currency, get_default_value
 from .account import Account
 from .profile import UserProfile
 from .transaction import TransactionJournal, Transaction
-from .categories import Category
+from .category import Category
